pyproject.toml
[tool.poetry]
name = "nonebot_plugin_memes_api"
version = "0.5.0b5"
description = "Nonebot2 plugin for making memes"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-memes-api"
repository = "https://github.com/noneplugin/nonebot-plugin-memes-api"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-plugin-alconna = ">=0.46.3,<1.0.0"
nonebot-plugin-uninfo = ">=0.6.1,<1.0.0"
nonebot-plugin-localstore = ">=0.6.0,<1.0.0"
nonebot-plugin-orm = ">=0.7.0,<1.0.0"
httpx = ">=0.20.0,<1.0.0"
pyyaml = "^6.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["fastapi"], version = "^2.3.0" }
nonebot-plugin-orm = { extras = ["default"], version = ">=0.7.0,<1.0.0" }
nonebot-adapter-onebot = "^2.2.0"

[tool.nonebot]
plugins = ["nonebot_plugin_memes_api"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
